Move Surface and friends to flow/
Circular dependencies are cropping up between shell/common/ and flow/,
so just move all of this stuff down into flow itself.  These classes are
all rasterization-related anyways.

shell/common/ has Dart dependenices, and a future CL will create
non-product and product permutations of any target with Dart
dependencies.  Moving all of the rasterization-related code to its own
library with no Dart dependencies means no permutations need to be made
of it.

Tested: Just compile, this is a mechanical change
