+++ flow/surface.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/common/surface.h"
+#include "flutter/flow/surface.h"
 
 namespace flutter {
