+++ flow/surface.h
@@ -2,15 +2,15 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_SHELL_COMMON_SURFACE_H_
-#define FLUTTER_SHELL_COMMON_SURFACE_H_
+#ifndef FLUTTER_FLOW_SURFACE_H_
+#define FLUTTER_FLOW_SURFACE_H_
 
 #include <memory>
 
 #include "flutter/flow/compositor_context.h"
 #include "flutter/flow/embedded_views.h"
+#include "flutter/flow/surface_frame.h"
 #include "flutter/fml/macros.h"
-#include "flutter/shell/common/surface_frame.h"
 
 namespace flutter {
 
@@ -39,4 +39,4 @@ class Surface {
 
 }  // namespace flutter
 
-#endif  // FLUTTER_SHELL_COMMON_SURFACE_H_
+#endif  // FLUTTER_FLOW_SURFACE_H_
