+++ flow/surface_frame.h
@@ -2,20 +2,20 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_SHELL_COMMON_SURFACE_FRAME_H_
-#define FLUTTER_SHELL_COMMON_SURFACE_FRAME_H_
+#ifndef FLUTTER_FLOW_SURFACE_FRAME_H_
+#define FLUTTER_FLOW_SURFACE_FRAME_H_
 
 #include <memory>
 
+#include "flutter/flow/gl_context_switch.h"
 #include "flutter/fml/macros.h"
-#include "flutter/shell/common/gl_context_switch.h"
 #include "third_party/skia/include/core/SkCanvas.h"
 #include "third_party/skia/include/core/SkSurface.h"
 
 namespace flutter {
 
-/// Represents a Frame that has been fully configured for the underlying client
-/// rendering API. A frame may only be submitted once.
+// This class represents a frame that has been fully configured for the
+// underlying client rendering API. A frame may only be submitted once.
 class SurfaceFrame {
  public:
   using SubmitCallback =
@@ -56,4 +56,4 @@ class SurfaceFrame {
 
 }  // namespace flutter
 
-#endif  // FLUTTER_SHELL_COMMON_SURFACE_FRAME_H_
+#endif  // FLUTTER_FLOW_SURFACE_FRAME_H_
