+++ shell/gpu/BUILD.gn
@@ -11,7 +11,6 @@ gpu_common_deps = [
   "//flutter/flow",
   "//flutter/fml",
   "//flutter/shell/common",
-  "//flutter/shell/common:common_standalone",
   "//third_party/skia",
 ]
