+++ shell/gpu/gpu_surface_gl.h
@@ -9,10 +9,10 @@
 #include <memory>
 
 #include "flutter/flow/embedded_views.h"
+#include "flutter/flow/gl_context_switch.h"
+#include "flutter/flow/surface.h"
 #include "flutter/fml/macros.h"
 #include "flutter/fml/memory/weak_ptr.h"
-#include "flutter/shell/common/gl_context_switch.h"
-#include "flutter/shell/common/surface.h"
 #include "flutter/shell/gpu/gpu_surface_gl_delegate.h"
 #include "third_party/skia/include/gpu/GrContext.h"
