+++ shell/platform/darwin/ios/ios_surface.h
@@ -10,9 +10,9 @@
 #include <memory>
 
 #include "flutter/flow/embedded_views.h"
+#include "flutter/flow/surface.h"
 #include "flutter/fml/macros.h"
 #include "flutter/fml/platform/darwin/scoped_nsobject.h"
-#include "flutter/shell/common/surface.h"
 
 @class CALayer;
