+++ shell/platform/fuchsia/flutter/BUILD.gn
@@ -315,6 +315,7 @@ executable("flutter_runner_unittests") {
     "//build/fuchsia/pkg:async-loop-default",
     "//build/fuchsia/pkg:scenic_cpp",
     "//build/fuchsia/pkg:sys_cpp_testing",
+    "//flutter/flow",
     "//flutter/lib/ui",
     "//flutter/runtime",
     "//flutter/shell/common",
