+++ shell/platform/fuchsia/flutter/surface.h
@@ -5,9 +5,9 @@
 #pragma once
 
 #include "compositor_context.h"
+#include "flutter/flow/surface.h"
 #include "flutter/fml/macros.h"
 #include "flutter/fml/memory/weak_ptr.h"
-#include "flutter/shell/common/surface.h"
 
 namespace flutter_runner {
